+++ src/components/PostCard/PostCard.tsx
@@ -1,13 +1,15 @@
-import styles from "./PostsCard.module.scss";
+import styles from "./PostCard.module.scss";
 import megaphoneIcon from "@/assets/megaphone.png";
 import dotsIcon from "@/assets/dots.png";
 import starIcon from "@/assets/star.png";
+import cn from "classnames";
+import { Categories } from "@/types/enums/Categories";
 
 interface Props {
   img: string;
   title: string;
   type: string;
-  filter: string;
+  category: Categories;
 }
 
 const getBorderStyle = (type: string): string => {
@@ -25,13 +27,13 @@ const getBorderStyle = (type: string): string => {
   }
 };
 
-const getIcon = (filter: string): string => {
-  switch (filter) {
-    case "Noticias":
+const getIcon = (category: Categories): string => {
+  switch (category) {
+    case Categories.Noticias:
       return megaphoneIcon;
-    case "Reseñas":
+    case Categories.Reseñas:
       return dotsIcon;
-    case "Eventos":
+    case Categories.Eventos:
       return starIcon;
     default:
       return "";
@@ -59,19 +61,23 @@ const PostCard = ({ img, title, type, category }: Props) => {
       <img
         src={img}
         alt={title}
-        className={`${styles.postsImage} ${getBorderStyle(type)}`}
+        className={cn(styles.postImage, getBorderStyle(type))}
       />
-      <div className={styles.postsTitleContainer}>
-        <p className={styles.postsTitle}>{title}</p>
+      <div className={styles.postTitleContainer}>
+        <p className={styles.postTitle}>{title}</p>
       </div>
-      <div className={styles.postsTypeContainer}>
-        <span className={`${styles.postsType} ${getTextColorStyle(type)}`}>
+      <div className={styles.postTypeContainer}>
+        <span className={cn(styles.postType, getTextColorStyle(type))}>
           {type.toUpperCase()}
         </span>
-        <img src={getIcon(filter)} alt={filter} className={styles.postsIcon} />
+        <img
+          src={getIcon(category)}
+          alt={category}
+          className={styles.postIcon}
+        />
       </div>
     </div>
   );
 };
 
-export default PostsCard;
+export default PostCard;
